requirements.txt
Django>=4.2,<5.0
python-dotenv==1.0.0
psycopg2-binary
boto3==1.28.57
Pillow
gunicorn
django-anymail[amazon_ses]
django-filter
beautifulsoup4~=4.13.4
openai~=1.78.0
pytz~=2025.2
django-storages
django-simple-history~=3.8.0
requests==2.31.0
urllib3<2.5.0
lockfile
django-parler~=2.3
django-simple-captcha~=0.6.2
yt-dlp==2025.7.21
azure-cognitiveservices-speech==1.42.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
xlsxwriter

# Cache y Redis
redis==5.0.1
django-redis==5.4.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
kombu==5.3.4

python-dateutil~=2.9.0.post0
Markdown~=3.8.2
selenium==4.26.1
webdriver-manager~=4.0.2
anthropic~=0.57.1
pandas~=2.3.1

# Error tracking and monitoring
sentry-sdk[django]

